cleaning more clang warnings

+++ src/core/cmt/setForwardModel.c
@@ -29,21 +29,21 @@
  *        observation, the 3*l+1'th row corresponds to the east observation,
  *        and 3*l+2'th row corresponds to the negative vertical observation.
  *
- * @param[in] l1          length of arrays
- * @param[in] ldevi       if true then compute the Green's functions matrix
- *                        G with the deviatoric constraint.
- *                        otherwise compute the general Green's functions
- *                        matrix for all six moment tensor terms.
- * @param[in] x1          x (North) receiver-source distance (meters) [l1]
- * @param[in] y1          y (East) receiver-source distance (meters) [l1]
- * @param[in] z1          z receiver-source distance (meters) [l1].
- *                        Note, that z increases up from the free
- *                        surface in the observation frame, hence, for most
- *                        applications z will be negative.
+ * @param[in] l1           length of arrays
+ * @param[in] ldeviatoric  if true then compute the Green's functions matrix
+ *                         G with the deviatoric constraint.
+ *                         otherwise compute the general Green's functions
+ *                         matrix for all six moment tensor terms.
+ * @param[in] x1           x (North) receiver-source distance (meters) [l1]
+ * @param[in] y1           y (East) receiver-source distance (meters) [l1]
+ * @param[in] z1           z receiver-source distance (meters) [l1].
+ *                         Note, that z increases up from the free
+ *                         surface in the observation frame, hence, for most
+ *                         applications z will be negative.
  *
- * @param[out] G          matrix of Green's functions stored in row major
- *                        format [3*l1 x ncol].  if ldevi is true then
- *                        ncol is 5, otherwise, ncol is 6.
+ * @param[out] G           matrix of Green's functions stored in row major
+ *                         format [3*l1 x ncol].  if ldeviatoric is true then
+ *                         ncol is 5, otherwise, ncol is 6.
  *
  * @result 0 indicates success
  *
@@ -143,7 +143,7 @@ int core_cmt_setForwardModel(const int l1, const bool ldeviatoric,
     // General case
     else
     {
-        memset(G, 0, 6*l1*sizeof(double));
+        memset(G, 0, (unsigned long) (6*l1)*sizeof(double));
         return -1;
     }
     return 0;
